Crate embedded_gui

Re-exports

pub use motion as animation;

pub use motion as animation_timeline;

pub use motion as animation_timing;

pub use motion as cinematic;

pub use motion as screen_transition;

pub use motion as transition_preset;

pub use motion as widget_animation;

pub use geometry::FluentBuilder;

pub use text::StringArena;

pub use text::TextSlice;

pub use view::FlexBuilder;

pub use view::Render;

pub use view::ViewContext;

pub use haptics::HapticPattern;

pub use haptics::HapticSequencer;

pub use i18n::LanguageId;

pub use i18n::TranslationEntry;

pub use i18n::TranslationTable;

pub use visual_widgets::BusyWheel;

pub use visual_widgets::GaugeWidget;

pub use visual_widgets::SpectrumAnalyzerWidget;

pub use visual_widgets::TouchInputFilter;

pub use adapter::ColorConvertedDrawTarget;

pub use adapter::DrawTargetColorExt;

pub use animation::Animation;

pub use animation::AnimationError;

pub use animation::AnimationHandlers;

pub use animation::AnimationId;

pub use animation::AnimationManager;

pub use animation::AnimationManagerCallbacks;

pub use animation::AnimationState;

pub use animation::Easing;

pub use animation::InertiaAnimator;

pub use animation::PathAnimator;

pub use animation::PathPoint;

pub use animation::RepeatMode;

pub use animation::SpringAnimator;

pub use animation::Timer;

pub use animation::Tween;

pub use animation::apply_easing;

pub use animation_timeline::AnimationGroup;

pub use animation_timeline::AnimationSequence;

pub use animation_timeline::ComposedAnimation;

pub use animation_timeline::ComposedAnimationCallbacks;

pub use animation_timeline::ComposedAnimationPlayer;

pub use animation_timeline::ComposedAnimationStatus;

pub use animation_timeline::CompositionControls;

pub use animation_timeline::CompositionMode;

pub use animation_timeline::Keyframe;

pub use animation_timeline::KeyframeTrack;

pub use animation_timeline::KeyframeTrackCallbacks;

pub use animation_timeline::SequencePlayer;

pub use animation_timeline::SequencePlayerStatus;

pub use animation_timeline::SequenceRepeatMode;

pub use animation_timeline::TimelineError;

pub use animation_timeline::TimelineStep;

pub use animation_timing::DEFAULT_DURATION_MS;

pub use animation_timing::FRAME_INTERVAL_MS;

pub use animation_timing::MOOOK_DURATION_MS;

pub use animation_timing::NORMALIZED_MAX;

pub use animation_timing::PORT_HOLE_DURATION_MS;

pub use animation_timing::SHUTTER_DURATION_MS;

pub use animation_timing::interpolate_moook;

pub use animation_timing::moook_curve;

pub use animation_timing::moook_duration_ms;

pub use animation_timing::timing_half_phase;

pub use animation_timing::timing_scaled;

pub use animation_timing::timing_shutter_phase;

pub use colors::ColorOps;

pub use gradient::ColorStop;

pub use gradient::MultiStopGradient;

pub use motion::stopwatch::PlaybackMode;

pub use motion::stopwatch::Stopwatch;

pub use nine_slice::BorderRect;

pub use nine_slice::NineSlice;

pub use nine_slice::NineSliceLayout;

pub use nine_slice::SlicePair;

pub use nine_slice::SliceScaleMode;

pub use power::PowerConfig;

pub use power::PowerEvent;

pub use power::PowerManager;

pub use power::PowerState;

pub use profiler::DirtyRectVisualizer;

pub use profiler::FrameBudgetTracker;

pub use quantize::BAYER_2X2;

pub use quantize::BAYER_4X4;

pub use quantize::DirtyBandAccumulator;

pub use quantize::bayer_1bit;

pub use quantize::bayer_2bit;

pub use quantize::bayer_4bit;

pub use quantize::bayer_dither_rgb565;

pub use spline::CatmullRomSpline;

pub use widgets::segmented::SegmentStyle;

pub use widgets::segmented::SevenSegmentDisplay;

pub use widgets::segmented::encode_7seg;

pub use widgets::segmented::encode_14seg;

pub use block::Block;

pub use cinematic::CardDeckDirection;

pub use cinematic::CardDeckState;

pub use cinematic::CardStory;

pub use cinematic::CardStoryTransition;

pub use cinematic::CinematicPreset;

pub use cinematic::GlanceTileSpec;

pub use cinematic::MotionTokens;

pub use cinematic::PeekRevealSpec;

pub use cinematic::TimelineMotionPreset;

pub use cinematic::animate_glance_focus;

pub use cinematic::animate_peek_reveal;

pub use cinematic::apply_carddeck_visibility;

pub use cinematic::setup_card_story;

pub use cinematic::setup_launcher_glance;

pub use cinematic::setup_launcher_glance_with_tokens;

pub use cinematic::setup_peek_timeline;

pub use cinematic::setup_peek_timeline_with_tokens;

pub use completion::CompletionSlot;

pub use completion::WaitTransfer;

pub use completion::WaitTransferFuture;

pub use context::GuiContext;

pub use context::GuiError;

pub use context::KeyBindingAction;

pub use context::PressTiming;

pub use context::WidgetBuilder;

pub use context::WidgetKeyBindings;

pub use context::WidgetKeyInputPolicy;

pub use display_backend::AsyncDmaTransfer;

pub use display_backend::DisplayBackend;

pub use display_backend::DisplayError;

pub use display_backend::DisplayRegion;

pub use display_backend::DmaTransfer;

pub use display_backend::SimulatorBackend;

pub use display_backend::TransferError;

pub use embassy::EmbassyWaitTransfer;

pub use embassy::EmbassyWaitTransferFuture;

pub use embassy::FrameClock;

pub use font::BitmapFont;

pub use font::Font;

pub use font::FontId;

pub use font::PackedFont;

pub use framebuffer::Framebuffer;

pub use framebuffer::FramebufferGray8;

pub use framebuffer::FramebufferRgba8888;

pub use framebuffer::FramebufferSlice;

pub use framebuffer::Rgba8888;

pub use geometry::Anchor;

pub use geometry::DirtyTracker;

pub use geometry::EdgeInsets;

pub use geometry::HorizontalAlign;

pub use geometry::Rect;

pub use image::BasicImageDecoder;

pub use image::EncodedImageFormat;

pub use image::ImageDecodeError;

pub use image::ImageDecoder;

pub use image::decode_image_auto;

pub use image::decode_image_with;

pub use image::decode_ppm_ascii;

pub use image::ImageAtlas;

pub use image::ImageAtlasEntry;

pub use image::ImageFit;

pub use image::ImageRef;

pub use image::ReelFrame;

pub use image::ReelPlayer;

pub use image::SpriteSheet;

pub use image::TileMode;

pub use image::TileRef;

pub use input::EventPhaseMask;

pub use input::InputEvent;

pub use input::NavDirection;

pub use input::PointerButton;

pub use input::PointerState;

pub use input::UiEvent;

pub use input::UiEventFilter;

pub use input::WidgetDispatchPolicy;

pub use input::WidgetEvent;

pub use input::WidgetEventFilter;

pub use input::WidgetEventKind;

pub use layout::Align;

pub use layout::Axis;

pub use layout::Constraint;

pub use layout::GridLayout;

pub use layout::GridPlacement;

pub use layout::GridTrack;

pub use layout::JustifyContent;

pub use layout::LayoutItem;

pub use layout::Length;

pub use layout::LinearLayout;

pub use mono::IconAlign;

pub use mono::IconPart;

pub use mono::MonoBitmap;

pub use palette::DisplayMode;

pub use palette::DisplayPalette;

pub use palette::InkRole;

pub use palette::RoleColors;

pub use pdc::PdcCommand;

pub use pdc::PdcCommandType;

pub use pdc::PdcImage;

pub use pdc::PdcPrecisePoint;

pub use present::PresentRegion;

pub use render::AntiAliasMode;

pub use render::Blend;

pub use render::BlendMode;

pub use render::CHAR_HEIGHT;

pub use render::CHAR_WIDTH;

pub use render::ColorFormat;

pub use render::Compositor;

pub use render::Dither;

pub use render::DrawTask;

pub use render::DrawTaskQueue;

pub use render::DrawUnit;

pub use render::EllipsisMode;

pub use render::Hardware2DAccelerator;

pub use render::LayerState;

pub use render::PartialBandBuffer;

pub use render::PathVerb;

pub use render::RenderBackendCaps;

pub use render::RenderCtx;

pub use render::RenderQuality;

pub use render::Software2DAccelerator;

pub use render::SoftwareDrawUnit;

pub use render::StrokeCap;

pub use render::StrokeDash;

pub use render::StrokeJoin;

pub use render::StrokeStyle;

pub use render::TextAlign;

pub use render::TextMetrics;

pub use render::TextOverflow;

pub use render::TextOverflowPolicy;

pub use render::TextStyle;

pub use render::TextWrap;

pub use render::Transform2D;

pub use render::VectorPath;

pub use render::VerticalAlign;

pub use render::dispatch_draw_tasks;

pub use render::LineBufferRenderer;

pub use render::ScanlineTarget;

pub use round::UnobstructedArea;

pub use round::circle_chord_width;

pub use round::round_screen_line_bounds;

pub use screen::Screen;

pub use screen::ScreenCommand;

pub use screen::ScreenId;

pub use screen::ScreenLifecycleEvent;

pub use screen::ScreenStack;

pub use screen::ScreenStackError;

pub use screen::ScreenTransition;

pub use screen_transition::ActiveScreenTransition;

pub use screen_transition::ScreenTransitionEffect;

pub use screen_transition::ScreenTransitionOrigin;

pub use screen_transition::ScreenTransitionRunner;

pub use screen_transition::ScreenTransitionSample;

pub use screen_transition::ScreenTransitionSpec;

pub use screen_transition::composite_framebuffer_fade;

pub use screen_transition::composite_framebuffer_scaled_y;

pub use screen_transition::fade_outgoing_opacity;

pub use screen_transition::render_transition_pair;

pub use state::CallbackSlot;

pub use state::FeedTimelineState;

pub use state::GuiModel;

pub use state::ListState;

pub use state::ModelChange;

pub use state::PropertySignal;

pub use state::ScrollState;

pub use state::Signal;

pub use state::SliceModel;

pub use state::SliderState;

pub use state::StateTransition;

pub use state::TabsState;

pub use state::WidgetStateMachine;

pub use style::AlphaLinearGradient;

pub use style::AlphaRadialGradient;

pub use style::Border;

pub use style::GradientDirection;

pub use style::LinearGradient;

pub use style::MultiPartStyle;

pub use style::PartStyleRule;

pub use style::Shadow;

pub use style::StateStyle;

pub use style::Style;

pub use style::StyleTransition;

pub use style::Theme;

pub use style::VisualState;

pub use style::VisualStateMask;

pub use style::WidgetPart;

pub use style::WidgetStyle;

pub use style::lerp_style;

pub use swapchain::StandardSwapChain;

pub use swapchain::SwapChain;

pub use swapchain::StandardTripleSwapChain;

pub use swapchain::TripleSwapChain;

pub use test_buffer::LayerCanvas;

pub use test_buffer::TestBuffer;

pub use text::BasicTextShaper;

pub use text::Line;

pub use text::ShapedGlyph;

pub use text::ShapingConfig;

pub use text::Span;

pub use text::Text;

pub use text::TextDirection;

pub use text::TextShaper;

pub use transition_preset::TransitionPreset;

pub use widget::EventContext;

pub use widget::EventPhase;

pub use widget::EventPolicy;

pub use widget::FocusGroupId;

pub use widget::MenuContract;

pub use widget::PropertyError;

pub use widget::PropertyKey;

pub use widget::PropertyValue;

pub use widget::StatefulWidget;

pub use widget::StyleClassId;

pub use widget::Widget;

pub use widget::WidgetFlags;

pub use widget::WidgetId;

pub use widget_animation::presets;

pub use widget_animation::AnimatedProperty;

pub use widget_animation::AnimationConflictPolicy;

pub use widget_animation::BindingSnapshot;

pub use widget_animation::WidgetAnimationCallbacks;

pub use widget_animation::WidgetAnimationError;

pub use widget_animation::WidgetAnimator;

pub use widget_animation::WidgetKeyframeState;

pub use widget_animation::WidgetPropertyKeyframe;

pub use widgets::ActionBarSlot;

pub use widgets::ActionBarWidget;

pub use widgets::ActionMenuError;

pub use widgets::ActionMenuItem;

pub use widgets::ActionMenuWidget;

pub use widgets::ActionableDialogWidget;

pub use widgets::ArcGaugeWidget;

pub use widgets::BandMeterWidget;

pub use widgets::BatteryState;

pub use widgets::ConfirmationDialogWidget;

pub use widgets::ContentIndicatorDirection;

pub use widgets::ContentIndicatorWidget;

pub use widgets::CrumbsIndicatorWidget;

pub use widgets::DialogAction;

pub use widgets::DialogError;

pub use widgets::DialogType;

pub use widgets::GaugeThreshold;

pub use widgets::InverterWidget;

pub use widgets::MatrixWidget;

pub use widgets::MeterOrientation;

pub use widgets::NotificationAction;

pub use widgets::NotificationError;

pub use widgets::NotificationPriority;

pub use widgets::NotificationSheetWidget;

pub use widgets::NumberPickerWidget;

pub use widgets::PeekBannerWidget;

pub use widgets::PickerError;

pub use widgets::ProgressBarWidget;

pub use widgets::RepeaterWidget;

pub use widgets::RichTextError;

pub use widgets::RichTextNodeWidget;

pub use widgets::ScaleMode;

pub use widgets::ScaleWidget;

pub use widgets::SelectionWidget;

pub use widgets::SparklineWidget;

pub use widgets::SpinboxWidget;

pub use widgets::StatusBarError;

pub use widgets::StatusBarMode;

pub use widgets::StatusBarWidget;

pub use widgets::TableWidget;

pub use widgets::TextSpan;

pub use widgets::TimeFormat;

pub use widgets::TimePickerField;

pub use widgets::TimePickerWidget;

pub use widgets::TimelineNodeState;

pub use widgets::TimelineNodeWidget;

pub use widgets::CarouselSpec;

pub use widgets::ChartMode;

pub use widgets::CompositeIconSpec;

pub use widgets::KeyboardLayout;

pub use widgets::NotificationLevel;

pub use widgets::SurfaceState;

pub use widgets::WidgetKind;

pub use widgets::WidgetNode;

Modules

adapter

Output boundary color conversion adapters for [DrawTarget].

block

colors

Standard CSS color constants and RGB565 conversion utilities.

completion

ISR-safe DMA completion signaling for async and sync present paths.

context

display_backend

Display backend abstraction for DMA-based rendering.

embassy

Optional Embassy integration helpers (feature embassy).

font

framebuffer

A RAM-backed, readback-capable framebuffer.

geometry

gradient

Multi-stop linear gradient rasterizer adapted from bevy_ui::gradients.

haptics

Lightweight haptic pattern sequencer for microcontrollers and embedded GUIs.

i18n

Zero-allocation internationalization (i18n) and localization for #![no_std] embedded targets. Allows microcontrollers to store compact string translation tables in Flash and switch languages dynamically at runtime without heap allocation.

image

input

interop

Optional bridges to the wider embedded-graphics ecosystem.

layout

mono

1-bit-per-pixel bitmaps and the composite icons built from them.

motion

nine_slice

9-Patch / 9-Slice scaling engine adapted from bevy_sprite::texture_slice.

palette

Semantic color roles and display modes for HUD / instrument panels.

pdc

Precise Draw Commands (PDC) vector graphics format and renderer.

power

Power management, inactivity timeouts, and dynamic refresh rate throttling.

prelude

present

profiler

UI performance profiling, frame pacing, and CPU load estimation for #![no_std] targets.

quantize

Low-color quantization, Bayer ordered dithering, and dirty band optimization.

render

round

Circular / Round Screen Geometry and Unobstructed Area adaptation.

screen

spline

Catmull-Rom and Cardinal cubic spline curves adapted from bevy_curve::cubic_splines.

state

style

swapchain

Swap chain implementation for double-buffered rendering.

test_buffer

text

view

visual_widgets

Custom visual widgets & DSP-accelerated UI controls for embedded-gui.

widget

widgets

Macros

gui_kdl

Compiles an inline KDL string literal into typed Rust code.

include_gui

Compiles an external KDL UI markup file into typed, zero-allocation Rust code.

Structs

EndianCorrectedBuffer

A backend for FrameBuf which changes the underlying byte order. This can be useful when using the buffer for DMA with peripherals that have a different endianness than the host.

FrameBuf

Constructs a frame buffer in memory. Lets you define the width(X), height (Y) and pixel type your using in your display (RGB, Monochrome etc.)

Enums

EndianCorrection

Enum indicating how the bytes should be converted in the host’s memory.

Traits

DMACapableFrameBufferBackend

Backends implementing this Trait can be used for DMA.

PixelRead

Re-exported from [embedded_draw_target] so that a buffer implementing it once is accepted by every crate in the ecosystem that needs readback, rather than by this one alone. Readback capability for a [DrawTarget].

Type Aliases

ActiveColor

Canonical pixel color type for the active target configuration.

CanvasColor

Alias for ActiveColor.

Pixel

Alias for ActiveColor.