embedded_gui::geometry

Trait FluentBuilder

pub trait FluentBuilder {
    // Provided methods
    fn map<U>(self, f: impl FnOnce(Self) -> U) -> U
       where Self: Sized { ... }
    fn when(self, condition: bool, then: impl FnOnce(Self) -> Self) -> Self
       where Self: Sized { ... }
    fn when_else(
        self,
        condition: bool,
        then: impl FnOnce(Self) -> Self,
        else_fn: impl FnOnce(Self) -> Self,
    ) -> Self
       where Self: Sized { ... }
    fn when_some<T>(
        self,
        option: Option<T>,
        then: impl FnOnce(Self, T) -> Self,
    ) -> Self
       where Self: Sized { ... }
    fn when_none<T>(
        self,
        option: &Option<T>,
        then: impl FnOnce(Self) -> Self,
    ) -> Self
       where Self: Sized { ... }
}

Fluent combinator trait for conditional and mapping transformations on builders.

Provided Methods

fn map<U>(self, f: impl FnOnce(Self) -> U) -> U
where Self: Sized,

Imperatively transform self with the provided closure.

fn when(self, condition: bool, then: impl FnOnce(Self) -> Self) -> Self
where Self: Sized,

Conditionally transform self when condition is true.

fn when_else( self, condition: bool, then: impl FnOnce(Self) -> Self, else_fn: impl FnOnce(Self) -> Self, ) -> Self
where Self: Sized,

Conditionally transform self with then if condition is true, otherwise with else_fn.

fn when_some<T>( self, option: Option<T>, then: impl FnOnce(Self, T) -> Self, ) -> Self
where Self: Sized,

Conditionally unwrap and transform self if option is Some(val).

fn when_none<T>( self, option: &Option<T>, then: impl FnOnce(Self) -> Self, ) -> Self
where Self: Sized,

Conditionally transform self if option is None.

Dyn Compatibility

This trait is dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety".