embedded_gui

Module prelude

Re-exports

pub use crate::ActionBarSlot;

pub use crate::ActionBarWidget;

pub use crate::ActiveColor;

pub use crate::ActiveScreenTransition;

pub use crate::Align;

pub use crate::AlphaLinearGradient;

pub use crate::AlphaRadialGradient;

pub use crate::AnimatedProperty;

pub use crate::Animation;

pub use crate::AnimationConflictPolicy;

pub use crate::AnimationError;

pub use crate::AnimationGroup;

pub use crate::AnimationHandlers;

pub use crate::AnimationId;

pub use crate::AnimationManager;

pub use crate::AnimationManagerCallbacks;

pub use crate::AnimationSequence;

pub use crate::AnimationState;

pub use crate::AntiAliasMode;

pub use crate::ArcGaugeWidget;

pub use crate::Axis;

pub use crate::BandMeterWidget;

pub use crate::BasicTextShaper;

pub use crate::BindingSnapshot;

pub use crate::BitmapFont;

pub use crate::Blend;

pub use crate::BlendMode;

pub use crate::Block;

pub use crate::Border;

pub use crate::BorderRect;

pub use crate::CallbackSlot;

pub use crate::CanvasColor;

pub use crate::CardDeckDirection;

pub use crate::CardDeckState;

pub use crate::CardStory;

pub use crate::CardStoryTransition;

pub use crate::CarouselSpec;

pub use crate::CatmullRomSpline;

pub use crate::ChartMode;

pub use crate::CinematicPreset;

pub use crate::ColorConvertedDrawTarget;

pub use crate::ColorFormat;

pub use crate::ColorOps;

pub use crate::ColorStop;

pub use crate::CompassMode;

pub use crate::CompassWidget;

pub use crate::ComposedAnimation;

pub use crate::ComposedAnimationCallbacks;

pub use crate::ComposedAnimationPlayer;

pub use crate::ComposedAnimationStatus;

pub use crate::CompositeIconSpec;

pub use crate::CompositionControls;

pub use crate::CompositionMode;

pub use crate::Compositor;

pub use crate::Constraint;

pub use crate::DirtyBandAccumulator;

pub use crate::DirtyRectVisualizer;

pub use crate::DirtyTracker;

pub use crate::Dither;

pub use crate::DrawTargetColorExt;

pub use crate::Easing;

pub use crate::EdgeInsets;

pub use crate::EllipsisMode;

pub use crate::EventContext;

pub use crate::EventPhase;

pub use crate::EventPhaseMask;

pub use crate::EventPolicy;

pub use crate::FeedTimelineState;

pub use crate::FlexBuilder;

pub use crate::FluentBuilder;

pub use crate::FocusGroupId;

pub use crate::Font;

pub use crate::FontId;

pub use crate::FrameBudgetTracker;

pub use crate::Framebuffer;

pub use crate::FramebufferGray8;

pub use crate::FramebufferRgba8888;

pub use crate::FramebufferSlice;

pub use crate::GaugeThreshold;

pub use crate::GlanceTileSpec;

pub use crate::GradientDirection;

pub use crate::GridLayout;

pub use crate::GridPlacement;

pub use crate::GridTrack;

pub use crate::GuiContext;

pub use crate::GuiError;

pub use crate::GuiModel;

pub use crate::HapticPattern;

pub use crate::HapticSequencer;

pub use crate::Hardware2DAccelerator;

pub use crate::IconAlign;

pub use crate::IconPart;

pub use crate::ImageAtlas;

pub use crate::ImageAtlasEntry;

pub use crate::ImageFit;

pub use crate::ImageRef;

pub use crate::InertiaAnimator;

pub use crate::InputEvent;

pub use crate::InverterWidget;

pub use crate::KeyBindingAction;

pub use crate::KeyboardLayout;

pub use crate::Keyframe;

pub use crate::KeyframeTrack;

pub use crate::KeyframeTrackCallbacks;

pub use crate::LanguageId;

pub use crate::LayerState;

pub use crate::LayoutItem;

pub use crate::Length;

pub use crate::Line;

pub use crate::LineBufferRenderer;

pub use crate::LinearGradient;

pub use crate::LinearLayout;

pub use crate::ListState;

pub use crate::MatrixWidget;

pub use crate::MenuCell;

pub use crate::MenuContract;

pub use crate::MeterOrientation;

pub use crate::ModelChange;

pub use crate::MonoBitmap;

pub use crate::MotionTokens;

pub use crate::MultiStopGradient;

pub use crate::NavDirection;

pub use crate::NineSlice;

pub use crate::NineSliceLayout;

pub use crate::NotificationLevel;

pub use crate::PackedFont;

pub use crate::PathAnimator;

pub use crate::PathPoint;

pub use crate::PathVerb;

pub use crate::PeekRevealSpec;

pub use crate::PlaybackMode;

pub use crate::PointerButton;

pub use crate::PointerState;

pub use crate::PowerConfig;

pub use crate::PowerEvent;

pub use crate::PowerManager;

pub use crate::PowerState;

pub use crate::PresentRegion;

pub use crate::PressTiming;

pub use crate::ProgressBarWidget;

pub use crate::PropertySignal;

pub use crate::Rect;

pub use crate::ReelFrame;

pub use crate::ReelPlayer;

pub use crate::Render;

pub use crate::RenderBackendCaps;

pub use crate::RenderCtx;

pub use crate::RenderQuality;

pub use crate::RepeatMode;

pub use crate::RepeaterWidget;

pub use crate::Rgba8888;

pub use crate::ScaleMode;

pub use crate::ScaleWidget;

pub use crate::ScanlineTarget;

pub use crate::Screen;

pub use crate::ScreenCommand;

pub use crate::ScreenId;

pub use crate::ScreenLifecycleEvent;

pub use crate::ScreenStack;

pub use crate::ScreenStackError;

pub use crate::ScreenTransition;

pub use crate::ScreenTransitionEffect;

pub use crate::ScreenTransitionOrigin;

pub use crate::ScreenTransitionRunner;

pub use crate::ScreenTransitionSample;

pub use crate::ScreenTransitionSpec;

pub use crate::ScrollState;

pub use crate::SegmentStyle;

pub use crate::SequencePlayer;

pub use crate::SequencePlayerStatus;

pub use crate::SequenceRepeatMode;

pub use crate::SevenSegmentDisplay;

pub use crate::Shadow;

pub use crate::ShapedGlyph;

pub use crate::ShapingConfig;

pub use crate::Signal;

pub use crate::SliceModel;

pub use crate::SlicePair;

pub use crate::SliceScaleMode;

pub use crate::SliderState;

pub use crate::Software2DAccelerator;

pub use crate::Span;

pub use crate::SparklineWidget;

pub use crate::SpinboxWidget;

pub use crate::SpringAnimator;

pub use crate::SpriteSheet;

pub use crate::StateStyle;

pub use crate::StateTransition;

pub use crate::StatefulWidget;

pub use crate::Stopwatch;

pub use crate::StringArena;

pub use crate::StrokeCap;

pub use crate::StrokeDash;

pub use crate::StrokeJoin;

pub use crate::StrokeStyle;

pub use crate::Style;

pub use crate::StyleClassId;

pub use crate::StyleTransition;

pub use crate::SurfaceState;

pub use crate::TableWidget;

pub use crate::TabsState;

pub use crate::Text;

pub use crate::TextAlign;

pub use crate::TextDirection;

pub use crate::TextMetrics;

pub use crate::TextOverflow;

pub use crate::TextOverflowPolicy;

pub use crate::TextShaper;

pub use crate::TextSlice;

pub use crate::TextStyle;

pub use crate::TextWrap;

pub use crate::Theme;

pub use crate::TileMode;

pub use crate::TileRef;

pub use crate::TimelineError;

pub use crate::TimelineMotionPreset;

pub use crate::TimelineStep;

pub use crate::Timer;

pub use crate::Transform2D;

pub use crate::TransitionPreset;

pub use crate::TranslationEntry;

pub use crate::TranslationTable;

pub use crate::Tween;

pub use crate::UiEvent;

pub use crate::UiEventFilter;

pub use crate::VectorPath;

pub use crate::VerticalAlign;

pub use crate::ViewContext;

pub use crate::VisualState;

pub use crate::WidgetAnimationCallbacks;

pub use crate::WidgetAnimationError;

pub use crate::WidgetAnimator;

pub use crate::WidgetDispatchPolicy;

pub use crate::WidgetEvent;

pub use crate::WidgetEventFilter;

pub use crate::WidgetEventKind;

pub use crate::WidgetFlags;

pub use crate::WidgetId;

pub use crate::WidgetKeyBindings;

pub use crate::WidgetKeyInputPolicy;

pub use crate::WidgetKeyframeState;

pub use crate::WidgetKind;

pub use crate::WidgetNode;

pub use crate::WidgetPropertyKeyframe;

pub use crate::WidgetStateMachine;

pub use crate::WidgetStyle;

pub use crate::animate_glance_focus;

pub use crate::animate_peek_reveal;

pub use crate::apply_carddeck_visibility;

pub use crate::apply_easing;

pub use crate::bayer_1bit;

pub use crate::bayer_2bit;

pub use crate::bayer_4bit;

pub use crate::bayer_dither_rgb565;

pub use crate::colors;

pub use crate::encode_7seg;

pub use crate::encode_14seg;

pub use crate::lerp_style;

pub use crate::presets;

pub use crate::render_transition_pair;

pub use crate::setup_card_story;

pub use crate::setup_launcher_glance;

pub use crate::setup_launcher_glance_with_tokens;

pub use crate::setup_peek_timeline;

pub use crate::setup_peek_timeline_with_tokens;

pub use crate::BasicImageDecoder;

pub use crate::EncodedImageFormat;

pub use crate::ImageDecodeError;

pub use crate::ImageDecoder;

pub use crate::LayerCanvas;

pub use crate::TestBuffer;

pub use crate::decode_image_auto;

pub use crate::decode_image_with;

pub use crate::decode_ppm_ascii;

Macros

gui_kdl

Compiles an inline KDL string literal into typed Rust code.

include_gui

Compiles an external KDL UI markup file into typed, zero-allocation Rust code.

Structs

Rgb565

Re-exported so generated code can name colors without the consumer depending on embedded-graphics-core directly. Rgb565 color.

Traits

PixelRead

Re-exported from [embedded_draw_target] so that a buffer implementing it once is accepted by every crate in the ecosystem that needs readback, rather than by this one alone. Readback capability for a [DrawTarget].